+++ src/app/credExplorer/weights/weightsToEdgeEvaluator.js
@@ -2,7 +2,7 @@
 
 import type {Edge} from "../../../core/graph";
 import type {WeightedTypes} from "./weights";
-import type {EdgeEvaluator} from "../../../core/attribution/pagerank";
+import type {EdgeEvaluator} from "../../../analysis/pagerank";
 import {NodeTrie, EdgeTrie} from "../../../core/trie";
 
 export function weightsToEdgeEvaluator(weights: WeightedTypes): EdgeEvaluator {
